+++ src/community/components/form/choice-group/choice-group.stories.tsx
@@ -13,6 +13,11 @@ import { ChoiceGroupItemProps } from './choice-group.types';
 const meta: Meta<typeof ChoiceGroup> = {
   component: ChoiceGroup,
   title: 'Community/Form/ChoiceGroup',
+  parameters: {
+    status: {
+      type: ['deprecated', 'ExistsInTediReady'],
+    },
+  },
 };
 
 export default meta;